requirements.txt
amqp==2.2.2
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
django-celery-results==1.0.1
django-extensions==1.9.7
django-grappelli==2.10.1
django-nested-admin==3.0.21
enum34==1.1.6
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
kombu==4.1.0
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-monkey-business==1.0.0
pytz==2017.3
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
suds-jurko==0.6
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
